materialized: split binary crate into multiple files
This makes space for some tracing-related support functions.
